Migrate to PHPUnit for all tests so Jenkins CI can run tests.
To run locally:

1. (Once only) Install PEAR, and make sure it's up to date
2. (Once only) Using PEAR, install PHPUnit, and make sure it installs with no errors
3. phpunit -c alltests.xml 

There are differences to the way in which PHPUnit versus SimpleTest work. I had to make a few changes to get PHPUnit to work at all. State between tests is reset, so if your tests expect test 1 to update variables or state, test 2 can't see those changes. If you rely on that behavior, do it in setUp() or within every test.