CSSCodec::encodeCharacter now throws InvalidArgumentException (instead of Exception) when passed a character having a codepoint of zero. Changed the corresponding expected exception in CSSCodecTest::testEncodeZero (which fixes the test).
Added a check of the return from CSSCodec::_parseHex in decodeCharacter so that
invalid escape sequences are left unchanged in the input.

CSSCodec::_parseHex now declines to decode escape sequences which would result
in a character having a codepoint greater than 0x10FFFF - fixes test:
1) CSSCodecTest::testDoNotDecodeInvalidCodePoint
Failed asserting that two strings are equal.
--- Expected
+++ Actual
@@ @@
-'\abcdefg'
+'&#11259375;g'

Replaced CSSCodecTest::testDecodeEatNullChar with what I believe to be the
correct expectation for that input:
CSSCodecTest::testDoNotDecodeInvalidZeroCodePoint.